+++ backend/canisters/group/impl/src/updates/accept_p2p_swap.rs
@@ -5,7 +5,7 @@ use canister_tracing_macros::trace;
 use group_canister::accept_p2p_swap::{Response::*, *};
 use ic_cdk_macros::update;
 use icrc_ledger_types::icrc1::transfer::TransferError;
-use types::{AcceptSwapSuccess, Chat, MessageId, MessageIndex, UserId};
+use types::{AcceptSwapSuccess, Chat, MessageId, MessageIndex, P2PSwapLocation, UserId};
 
 #[update]
 #[trace]
@@ -78,7 +78,11 @@ fn reserve_p2p_swap(args: &Args, state: &mut RuntimeState) -> Result<ReserveP2PS
                     user_id,
                     c2c_args: user_canister::c2c_accept_p2p_swap::Args {
                         swap_id: result.content.swap_id,
-                        chat: Chat::Group(state.env.canister_id().into()),
+                        location: P2PSwapLocation::from_message(
+                            Chat::Group(state.env.canister_id().into()),
+                            args.thread_root_message_index,
+                            args.message_id,
+                        ),
                         created: result.created,
                         created_by: result.created_by,
                         token0: result.content.token0,