+++ frontend/openchat-client/src/openchat.ts
@@ -5159,6 +5159,13 @@ export class OpenChat extends OpenChatAgentWorker {
         });
     }
 
+    reportedMessages(userId: string | undefined): Promise<string> {
+        return this.sendRequest({
+            kind: "reportedMessages",
+            userId,
+        });
+    }
+
     payForDiamondMembership(
         token: string,
         duration: DiamondMembershipDuration,