+++ frontend/openchat-shared/src/domain/worker.ts
@@ -311,6 +311,7 @@ export type WorkerRequest =
     | ApproveTransfer
     | DeleteDirectChat
     | GetDiamondMembershipFees
+    | GetReportedMessages
     | SetTranslationCorrection
     | ApproveTranslationCorrection
     | RejectTranslationCorrection
@@ -1472,6 +1473,11 @@ type GetExchangeRates = {
     kind: "exchangeRates";
 };
 
+type GetReportedMessages = {
+    kind: "reportedMessages";
+    userId: string | undefined;
+};
+
 type AcceptP2PSwapOffer = {
     chatId: ChatIdentifier;
     threadRootMessageIndex: number | undefined;
@@ -1784,6 +1790,8 @@ export type WorkerResult<T> = T extends PinMessage
     ? boolean
     : T extends GetDiamondMembershipFees
     ? DiamondMembershipFees[]
+    : T extends GetReportedMessages
+    ? string
     : T extends GetExchangeRates
     ? Record<string, TokenExchangeRates>
     : T extends SetTranslationCorrection